refactor(backend-cuda): split — extract BackendCollective + BackendGraph (#8 Phase 2)
Move two more supertraits out of the 4765-line `cuda/mod.rs`:

- `cuda/collective.rs` (60 lines): `BackendCollective` impl —
  TP world_size/rank from FERRUM_TP/FERRUM_RANK env, all_reduce
  Sum-only stub, all_gather/broadcast placeholders.
- `cuda/graph.rs` (321 lines): `BackendGraph` impl + `GraphSlotRaw`
  + DECODE_GRAPHS multi-slot cache + helpers (graph_slots,
  install_decode_graph_raw, with_decode_graph,
  invalidate_decode_graph, invalidate_all_decode_graphs).

`decode_state_slot` stays in mod.rs (also used by core Backend ops);
visibility bumped to `pub(super)` so `cuda::graph` can reach it.

After Phase 2: mod.rs 4765 → 4408 lines (-7.5%). No behaviour
change — pure mechanical extraction.
