+++ crates/ferrum-kernels/src/backend/cuda/mod.rs
@@ -25,8 +25,12 @@
 // keeps `impl Backend for CudaBackend` (the core trait) plus the
 // supertraits whose impls haven't been extracted yet. Phase 1 split
 // (#8): INT8 KV (`BackendInt8KvOps` + helpers + `KvCacheQuant`
-// constructor) → `cuda/int8_kv.rs`. Future phases extract the larger
-// supertraits (Marlin, PagedKv, MoeFused, Graph, Collective).
+// constructor) → `cuda/int8_kv.rs`. Phase 2: BackendCollective →
+// `cuda/collective.rs`, BackendGraph + GraphSlotRaw + DECODE_GRAPHS
+// helpers → `cuda/graph.rs`. Future phases extract the remaining
+// supertraits (Marlin, PagedKv, MoeFused).
+pub mod collective;
+pub mod graph;
 pub mod int8_kv;
 pub use int8_kv::{OptionalCudaInt8, OptionalCudaScalesF16};
 
@@ -1659,256 +1663,6 @@ impl Backend for CudaBackend {
     }
 }
 
-impl BackendCollective for CudaBackend {
-    // ── TP collectives ──────────────────────────────────────────────────
-    //
-    // World size / rank come from env vars (FERRUM_TP, FERRUM_RANK).
-    // The NcclRank group itself is constructed by the executor (which
-    // has access to all GPU streams needed for `NcclRank::init_all`).
-    // all_reduce is wired through `crate::nccl_comm::NcclRank`; the other
-    // two are placeholders until `nccl_comm` gains them (they're not
-    // blocking on the LLM decode path — single-rank skips these entirely).
-
-    fn world_size(_ctx: &Self::Context) -> usize {
-        std::env::var("FERRUM_TP")
-            .ok()
-            .and_then(|s| s.parse::<usize>().ok())
-            .unwrap_or(1)
-    }
-
-    fn rank(_ctx: &Self::Context) -> usize {
-        std::env::var("FERRUM_RANK")
-            .ok()
-            .and_then(|s| s.parse::<usize>().ok())
-            .unwrap_or(0)
-    }
-
-    fn all_reduce(ctx: &mut Self::Context, buf: &mut Self::Buffer, len: usize, op: ReduceOp) {
-        // Only Sum is supported for now (the NCCL wrapper is sum-only).
-        if !matches!(op, ReduceOp::Sum) {
-            tracing::warn!(
-                "CudaBackend::all_reduce: op {op:?} not implemented (only Sum); skipping"
-            );
-            return;
-        }
-        // Single-rank path: no-op.
-        if Self::world_size(ctx) <= 1 {
-            return;
-        }
-        // Multi-rank path: requires the executor to have constructed a
-        // shared NcclRank and attached it to thread-local state. The
-        // current NcclRank API (`crate::nccl_comm::NcclRank::init_all`) is
-        // process-global and we don't want to reach into it from a
-        // Backend method. Leaving a runtime warning so misuse surfaces.
-        tracing::warn!(
-            "CudaBackend::all_reduce: FERRUM_TP > 1 but no NcclRank attached to \
-             CudaState — requires executor-level wiring (Phase E-TP)."
-        );
-    }
-
-    fn all_gather(
-        _ctx: &mut Self::Context,
-        _local: &Self::Buffer,
-        _global: &mut Self::Buffer,
-        _local_len: usize,
-    ) {
-        // Phase E-TP: no NCCL wrapper for all_gather yet.
-    }
-
-    fn broadcast(_ctx: &mut Self::Context, _buf: &mut Self::Buffer, _len: usize, _src_rank: usize) {
-        // Phase E-TP: no NCCL wrapper for broadcast yet.
-    }
-}
-
-impl BackendGraph for CudaBackend {
-    fn set_decode_state(ctx: &mut Self::Context, token: u32, step: u32) {
-        let valid_kv = (step as i32) + 1;
-        let step_i = step as i32;
-        let stream = ctx.stream.clone();
-        let mut w = decode_state_slot().write().expect("DECODE_STATE poisoned");
-        let bufs = w.as_mut().expect("DecodeStateBufs not initialised");
-        stream
-            .memcpy_htod(&[token], &mut bufs.token)
-            .expect("token_buf memcpy");
-        stream
-            .memcpy_htod(&[step_i], &mut bufs.pos)
-            .expect("pos_buf memcpy");
-        stream
-            .memcpy_htod(&[valid_kv], &mut bufs.kv)
-            .expect("kv_buf memcpy");
-    }
-
-    fn set_dev_state_mode(ctx: &mut Self::Context, enable: bool) {
-        ctx.use_dev_state = enable;
-    }
-
-    fn begin_graph_capture(ctx: &mut Self::Context) -> Result<()> {
-        use cudarc::driver::sys::CUstreamCaptureMode;
-        // Event tracking already disabled globally in default_stream; begin
-        // capture directly in relaxed mode. Bare-Rust cudarc reproducer
-        // confirms this configuration works on Blackwell + CUDA 13
-        // (`cudarc_graph_no_event_tracking` test). The full ferrum bench
-        // path still SIGSEGVs though — remaining delta is likely one of
-        // PTX module load timing, cuBLAS workspace interaction, or a
-        // specific kernel's use of constant memory that doesn't survive
-        // capture. See `docs/phase-e-cuda-status.md` graph section.
-        ctx.stream
-            .begin_capture(CUstreamCaptureMode::CU_STREAM_CAPTURE_MODE_RELAXED)
-            .map_err(|e| FerrumError::unsupported(format!("begin_capture: {e}")))?;
-        ctx.capture_in_flight = true;
-        Ok(())
-    }
-
-    fn end_graph_capture(ctx: &mut Self::Context, key: u64) -> Result<()> {
-        use cudarc::driver::sys;
-        if !ctx.capture_in_flight {
-            return Err(FerrumError::unsupported("end_capture without begin"));
-        }
-        ctx.capture_in_flight = false;
-
-        // Bypass cudarc's end_capture — it does cuStreamEndCapture +
-        // cuGraphInstantiateWithFlags in one call, and one of those corrupts
-        // the context on Blackwell. Call them separately so we can see which.
-        ctx.ctx
-            .bind_to_thread()
-            .map_err(|e| FerrumError::unsupported(format!("bind pre-end: {e}")))?;
-
-        let cu_stream = ctx.stream.cu_stream();
-        let mut cu_graph: sys::CUgraph = std::ptr::null_mut();
-        let st1 = unsafe { sys::cuStreamEndCapture(cu_stream, &mut cu_graph) };
-        if st1 != sys::CUresult::CUDA_SUCCESS {
-            return Err(FerrumError::unsupported(format!(
-                "cuStreamEndCapture failed: {st1:?}"
-            )));
-        }
-        if cu_graph.is_null() {
-            return Err(FerrumError::unsupported(
-                "cuStreamEndCapture returned null graph",
-            ));
-        }
-
-        // flags=0: no AUTO_FREE_ON_LAUNCH. The captured graph contains
-        // only kernel launches (memcpys are sync via cuMemcpyHtoD_v2
-        // outside capture, see populate_batched_pointers), so
-        // AUTO_FREE has nothing to free. With AUTO_FREE on, replays
-        // worked for ~14 iters then SIGSEGV — likely device-side
-        // launch resources getting freed mid-launch sequence.
-        let flags = 0u64;
-        let mut cu_graph_exec: sys::CUgraphExec = std::ptr::null_mut();
-        let st2 = unsafe { sys::cuGraphInstantiateWithFlags(&mut cu_graph_exec, cu_graph, flags) };
-        if st2 != sys::CUresult::CUDA_SUCCESS {
-            unsafe {
-                sys::cuGraphDestroy(cu_graph);
-            }
-            return Err(FerrumError::unsupported(format!(
-                "cuGraphInstantiate failed: {st2:?}"
-            )));
-        }
-
-        // Upload graph to GPU before first launch. Without this, the first
-        // cuGraphLaunch does lazy JIT + resource upload while the stream
-        // still has pending ops — libcuda dereferences not-yet-uploaded
-        // graph state and SIGSEGVs on Blackwell + CUDA 13.
-        let st3 = unsafe { sys::cuGraphUpload(cu_graph_exec, cu_stream) };
-        if st3 != sys::CUresult::CUDA_SUCCESS {
-            unsafe {
-                sys::cuGraphExecDestroy(cu_graph_exec);
-                sys::cuGraphDestroy(cu_graph);
-            }
-            return Err(FerrumError::unsupported(format!(
-                "cuGraphUpload failed: {st3:?}"
-            )));
-        }
-
-        // Install into the multi-slot cache keyed by `key`. Replaces any
-        // existing graph for the same key; the old GraphSlotRaw drops
-        // (cuCtxSync + cuGraphExecDestroy + cuGraphDestroy in its Drop
-        // impl) before the new one takes its place.
-        install_decode_graph_raw(key, cu_graph, cu_graph_exec, ctx.stream.clone());
-        Ok(())
-    }
-
-    fn reset_graph(_ctx: &mut Self::Context, key: u64) {
-        invalidate_decode_graph(key);
-    }
-
-    fn reset_all_graphs(_ctx: &mut Self::Context) {
-        invalidate_all_decode_graphs();
-    }
-
-    fn replay_graph(ctx: &mut Self::Context, key: u64) -> Result<bool> {
-        use cudarc::driver::sys;
-        let cu_stream = ctx.stream.cu_stream();
-        ctx.ctx
-            .bind_to_thread()
-            .map_err(|e| FerrumError::unsupported(format!("bind pre-replay: {e}")))?;
-        with_decode_graph(key, |g_opt| {
-            if let Some(g) = g_opt {
-                let prof = std::env::var("FERRUM_GRAPH_PROF").is_ok();
-                let t_pre = if prof {
-                    Some(std::time::Instant::now())
-                } else {
-                    None
-                };
-                // Re-upload before each launch. Without it, c=4 throughput
-                // drops 257→178 tok/s (post-Phase-8 measurement). The
-                // graph instantiate-then-upload-once design didn't pan out
-                // empirically; keep the per-replay upload until we
-                // understand why removing it slows things down.
-                let skip_upload =
-                    std::env::var("FERRUM_GRAPH_SKIP_UPLOAD").map_or(false, |v| v == "1");
-                if !skip_upload {
-                    let st_up = unsafe { sys::cuGraphUpload(g.cu_graph_exec, cu_stream) };
-                    if st_up != sys::CUresult::CUDA_SUCCESS {
-                        return Err(FerrumError::unsupported(format!(
-                            "cuGraphUpload: {st_up:?}"
-                        )));
-                    }
-                }
-                let t_after_upload = if prof {
-                    Some(std::time::Instant::now())
-                } else {
-                    None
-                };
-                let st = unsafe { sys::cuGraphLaunch(g.cu_graph_exec, cu_stream) };
-                if st != sys::CUresult::CUDA_SUCCESS {
-                    return Err(FerrumError::unsupported(format!("cuGraphLaunch: {st:?}")));
-                }
-                let t_after_launch = if prof {
-                    Some(std::time::Instant::now())
-                } else {
-                    None
-                };
-                let skip_sync = std::env::var("FERRUM_GRAPH_SKIP_SYNC").map_or(false, |v| v == "1");
-                if !skip_sync {
-                    let st_sync = unsafe { sys::cuStreamSynchronize(cu_stream) };
-                    if st_sync != sys::CUresult::CUDA_SUCCESS {
-                        return Err(FerrumError::unsupported(format!(
-                            "post-launch sync: {st_sync:?}"
-                        )));
-                    }
-                }
-                if let (Some(t0), Some(t1), Some(t2)) = (t_pre, t_after_upload, t_after_launch) {
-                    static N: std::sync::atomic::AtomicU64 = std::sync::atomic::AtomicU64::new(0);
-                    let n = N.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-                    if n.is_multiple_of(64) {
-                        let upload = t1.duration_since(t0).as_micros();
-                        let launch = t2.duration_since(t1).as_micros();
-                        let sync = t2.elapsed().as_micros();
-                        eprintln!(
-                            "[graph-prof] call#{n} upload={upload}us launch={launch}us sync={sync}us total={}us",
-                            t0.elapsed().as_micros()
-                        );
-                    }
-                }
-                Ok(true)
-            } else {
-                Ok(false)
-            }
-        })
-    }
-}
-
 // ────────────────────────────────────────────────────────────────────────
 // Process-global stream for context-free ops
 // ────────────────────────────────────────────────────────────────────────
@@ -1994,113 +1748,6 @@ pub fn install_thread_stream(stream: Arc<CudaStream>) {
     *stream_slot().write().expect("GLOBAL_STREAM poisoned") = Some(stream);
 }
 
-// ────────────────────────────────────────────────────────────────────────
-// Process-global decode graph slot
-// ────────────────────────────────────────────────────────────────────────
-//
-// Stored here (not on `CudaState`) because:
-//   - Backend::Context isn't Send+Sync for all backends (Metal holds a
-//     raw CommandBufferRef) — the model struct gets Send issues if ctx
-//     is stored on it.
-//   - Only CUDA uses graph capture, so global-per-process is fine.
-//   - Kernel arg pointers captured in the graph reference model-owned
-//     scratch buffers; the model outlives any graph, so no dangling refs.
-//
-// `CudaGraph` isn't automatically `Send+Sync` in cudarc's public API —
-// we wrap in our own marker struct with `unsafe impl`. The stream itself
-// is single-threaded per model (engine serialises via Mutex), so graph
-// launch from the same thread is safe.
-
-/// Raw graph slot holding cuGraph + cuGraphExec pointers directly, bypassing
-/// cudarc's CudaGraph wrapper. The wrapper's `end_capture` does
-/// cuStreamEndCapture + cuGraphInstantiateWithFlags in one non-overridable
-/// call, and one of those corrupts the context on Blackwell; bypassing lets
-/// us split the FFI calls and choose which instantiate variant to use.
-struct GraphSlotRaw {
-    cu_graph: cudarc::driver::sys::CUgraph,
-    cu_graph_exec: cudarc::driver::sys::CUgraphExec,
-    // Keep the stream Arc alive so its underlying cu_stream stays valid.
-    _stream: std::sync::Arc<cudarc::driver::CudaStream>,
-}
-
-impl Drop for GraphSlotRaw {
-    fn drop(&mut self) {
-        use cudarc::driver::sys;
-        unsafe {
-            // Sync device before destroying graph resources to ensure no
-            // kernel launches from this graph are still in flight.
-            sys::cuCtxSynchronize();
-            if !self.cu_graph_exec.is_null() {
-                sys::cuGraphExecDestroy(self.cu_graph_exec);
-            }
-            if !self.cu_graph.is_null() {
-                sys::cuGraphDestroy(self.cu_graph);
-            }
-            // Sync again after destroy so any cleanup completes.
-            sys::cuCtxSynchronize();
-        }
-    }
-}
-
-// SAFETY: graph launch is serialised through the model's stream, which
-// is accessed from one thread at a time (engine Mutex-wraps the model).
-unsafe impl Send for GraphSlotRaw {}
-unsafe impl Sync for GraphSlotRaw {}
-
-// Multi-slot graph cache, keyed by an opaque `u64`. Caller chooses the
-// key — the model uses `m_padded` (or 0 for single-item) so that
-// different batch shapes get their own captured graph instead of
-// thrashing a single slot at every shape change.
-//
-// Native CUDA microbench (graph_upload_bench.cu, 320 launches × 500 iters,
-// alternating two graph sizes) confirmed multi-slot replay is stable
-// at ~0.26ms/iter with no degradation vs single slot.
-static DECODE_GRAPHS: std::sync::OnceLock<std::sync::RwLock<HashMap<u64, GraphSlotRaw>>> =
-    std::sync::OnceLock::new();
-
-fn graph_slots() -> &'static std::sync::RwLock<HashMap<u64, GraphSlotRaw>> {
-    DECODE_GRAPHS.get_or_init(|| std::sync::RwLock::new(HashMap::new()))
-}
-
-fn install_decode_graph_raw(
-    key: u64,
-    cu_graph: cudarc::driver::sys::CUgraph,
-    cu_graph_exec: cudarc::driver::sys::CUgraphExec,
-    stream: std::sync::Arc<cudarc::driver::CudaStream>,
-) {
-    let mut g = graph_slots().write().expect("DECODE_GRAPHS poisoned");
-    g.insert(
-        key,
-        GraphSlotRaw {
-            cu_graph,
-            cu_graph_exec,
-            _stream: stream,
-        },
-    );
-}
-
-fn with_decode_graph<R>(key: u64, f: impl FnOnce(Option<&GraphSlotRaw>) -> Result<R>) -> Result<R> {
-    let guard = graph_slots().read().expect("DECODE_GRAPHS poisoned");
-    f(guard.get(&key))
-}
-
-/// Evict ONE cached graph — call when its kernel-arg pointers (KV cache,
-/// scratch buffers) might be invalidated.
-pub fn invalidate_decode_graph(key: u64) {
-    graph_slots()
-        .write()
-        .expect("DECODE_GRAPHS poisoned")
-        .remove(&key);
-}
-
-/// Evict ALL cached graphs — used by hard reset (model reload, scratch
-/// realloc) when every captured pointer might be stale.
-pub fn invalidate_all_decode_graphs() {
-    graph_slots()
-        .write()
-        .expect("DECODE_GRAPHS poisoned")
-        .clear();
-}
 
 // ────────────────────────────────────────────────────────────────────────
 // Process-global decode state buffers (token_id, pos, kv_len)
@@ -2122,7 +1769,7 @@ unsafe impl Sync for DecodeStateBufs {}
 static DECODE_STATE: std::sync::OnceLock<std::sync::RwLock<Option<DecodeStateBufs>>> =
     std::sync::OnceLock::new();
 
-fn decode_state_slot() -> &'static std::sync::RwLock<Option<DecodeStateBufs>> {
+pub(super) fn decode_state_slot() -> &'static std::sync::RwLock<Option<DecodeStateBufs>> {
     DECODE_STATE.get_or_init(|| std::sync::RwLock::new(None))
 }
